requirements.txt
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
cssselect2==0.7.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.7
docopt==0.6.2
docx2pdf==0.1.8
et-xmlfile==1.1.0
excel2img==1.4.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
flash==1.0.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
gingerit==0.0.0.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
img2pdf==0.5.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
json5==0.9.25
jsonify==0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
language-tool-python==2.8
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdf2docx==0.5.8
pdfkit==1.0.0
pikepdf==9.1.1
pillow==10.4.0
pillow_heif==0.18.0
pipenv==2024.0.1
pipwin==0.5.2
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycairo==1.26.1
pycparser==2.22
pydicom==2.4.4
pydyf==0.11.0
Pygments==2.18.0
pyjsparser==2.7.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypandoc==1.13
PyPDF2==3.0.1
pyphen==0.16.0
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
python-pptx==1.0.2
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scipy==1.14.0
Send2Trash==1.8.3
setuptools==70.3.0
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlparse==0.5.0
stack-data==0.6.3
surface==0.8.0
svg2png==1.2
svg2rlg==0.3
svglib==1.5.1
svgpathtools==1.6.1
svgwrite==1.4.3
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.26.3
Wand==0.6.13
wcwidth==0.2.13
weasyprint==62.3
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.13
wkhtmltopdf==0.2
wrapt==1.16.0
xlsx2html==0.6.1
XlsxWriter==3.2.0
zopfli==0.2.3
